Moved relevant tests from Site to SiteRepository and added tests for SiteRepository
