+++ Tests/Integration/Domain/Site/Fixtures/can_get_all_sites.xml
